jarretje

+++ src/main/java/nl/ronaldteune/ili9486/Tester.java
@@ -14,7 +14,7 @@ public class Tester {
         Thread.sleep(2000);
         final short[] data = new short[480];
         for (short i = 0; i < data.length; i++) {
-            data[i] = 12345 - i * 10;
+            data[i] = (short) (12345 - i * 10);
         }
         ili9486.drawRow(10, data);
